Moved bandcamp page classes one level up
