+++ src/bandcamp/page/schema.ts
